create baseindexingcommand for common functions
